pyproject.toml
[tool.poetry]
name = "dmail"
version = "0.1.0"
description = "Disposable mail server"
authors = ["pajecawav <pajecawav@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^1.1.2"
pyTelegramBotAPI = "^3.7.4"
python-dotenv = "^0.15.0"
Flask-SQLAlchemy = "^2.4.4"
APScheduler = "^3.6.3"
pyOpenSSL = "^20.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
